pyproject.toml
[tool.poetry]
name = "loansystem"
version = "0.1.0"
description = ""
authors = ["jggomez <juan.gomez01@gmail.com>"]
readme = "README.md"
packages = [
    { include = "backend", from = "src" },
    { include = "ui", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.12.5"
google-genai = "^1.55.0"
pyyaml = "^6.0.3"
google-cloud-firestore = "^2.21.0"
python-dotenv = "^1.2.1"
google-cloud-storage = "^3.7.0"
streamlit = "^1.52.1"
streamlit-pdf-viewer = "^0.0.26"
pymupdf = "^1.26.7"
scikit-learn = "^1.4.2"
numpy = ">=1.26.0,<2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
